+++ docs/CHANGELOG.rst
@@ -13,6 +13,7 @@ Unreleased
 Changed
 -------
 - **BACKWARD INCOMPATIBLE:** Drop support for ``Python 3.11``
+- Support dynamic calculation of processes' resource requirements.
 - Add ``process_exception`` method to auditlog middleware
 - Add ``deleted`` field to ``AnnotationValues`` model
 - Remove not used method ``remove_delete_markers`` from the ``AnnotationValue``